requirements.txt
torch
torchvision
albumentations
segmentation-models-pytorch
scikit-learn
pyyaml
tqdm
Pillow
